requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.20.26
botocore==1.23.54
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
colorama==0.4.6
constantly==15.1.0
crispy-bootstrap5==0.7
cryptography==41.0.2
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==0.5.0
Django==4.2.2
django-admin-sortable2==2.1.10
django-allauth==0.54.0
django-anymail==9.1
django-auto-logout==0.5.0
django-ckeditor==6.0.0
django-ckeditor-5==0.1.6
django-crispy-forms==2.0
django-dbbackup==4.0.2
django-dotenv==1.4.2
django-environ==0.9.0
django-extensions==3.2.3
django-filter==23.3
django-formset-js-improved==0.5.0.2
django-formtools==2.4.1
django-grappelli==3.0.8
django-heroku==0.3.1
django-htmx==1.14.0
django-import-export==3.2.0
django-jazzmin==2.6.0
django-jet==1.0.8
django-jquery-js==3.1.1
django-js-asset==1.2.2
django-mailgun-provider==0.2.3
django-mathfilters==1.0.0
django-on-heroku==1.1.2
django-otp==1.2.4
django-paypal==2.0
django-phonenumber-field==6.4.0
django-plaintext-password==0.1.0
django-recaptcha==3.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-six==1.0.5
django-social-share==2.2.1
django-static-fontawesome==5.14.0.0
django-storages==1.12.3
django-taggit==3.0.0
django-templated-mail==1.1.1
django-tinymce==3.4.0
django-two-factor-auth==1.15.4
django-user-agents==0.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.0.5
environs==9.5.0
et-xmlfile==1.1.0
fontawesomefree==6.4.0
frozenlist==1.4.0
geoip2==4.6.0
gunicorn==21.2.0
humanize==4.6.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==0.10.0
Markdown==3.4.3
MarkupPy==1.14
marshmallow==3.19.0
maxminddb==2.4.0
msgpack==1.0.7
multidict==6.0.4
numpy==1.26.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.1.1
paypal-payouts-sdk==1.0.0
paypalhttp==1.0.1
phonenumbers==8.13.18
Pillow==9.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.1.0
s3transfer==0.5.2
service-identity==23.1.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
stripe==6.6.0
tablib==3.5.0
twilio==8.5.0
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
ua-parser==0.18.0
urllib3==1.26.17
user-agents==2.2.0
validate-email==1.3
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
zope.interface==6.1
django-notifications-hf==1.8.0
django-pandas==0.6.4
plotly==5.17.0
reportlab==4.0.6
